Use T4 to generate Quarternion types.
