+++ Source/OpenTK/OpenTK.csproj
@@ -162,6 +162,11 @@
     <Compile Include="Math\Matrix4x2d.cs" />
     <Compile Include="Math\Matrix4x3.cs" />
     <Compile Include="Math\Matrix4x3d.cs" />
+    <Compile Include="Math\Quaternion.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>Quaternion.tt</DependentUpon>
+    </Compile>
     <Compile Include="Platform\Common\Hid.cs" />
     <Compile Include="Platform\DisplayDeviceBase.cs" />
     <Compile Include="Platform\Egl\EglUnixContext.cs" />
@@ -508,9 +513,6 @@
     <Compile Include="Input\IInputDevice.cs">
       <SubType>Code</SubType>
     </Compile>
-    <Compile Include="Math\Quaternion.cs">
-      <SubType>Code</SubType>
-    </Compile>
     <Compile Include="Math\Matrix3d.cs">
       <SubType>Code</SubType>
     </Compile>
@@ -563,9 +565,6 @@
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="Math\MathHelper.cs">
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="Math\Quaterniond.cs">
       <SubType>Code</SubType>
     </Compile>
     <Compile Include="Math\Vector4h.cs">
@@ -757,6 +756,10 @@
     <Compile Include="Input\HatPosition.cs" />
     <Compile Include="Input\JoystickHatState.cs" />
     <Compile Include="MouseCursor.cs" />
+    <None Include="Math\Quaternion.tt">
+      <Generator>TextTemplatingFileGenerator</Generator>
+      <LastGenOutput>Quaternion.cs</LastGenOutput>
+    </None>
     <Compile Include="Input\KeyModifiers.cs" />
     <Compile Include="Platform\MacOS\CocoaContext.cs" />
     <Compile Include="Platform\MacOS\CocoaNativeWindow.cs" />
